requirements.txt
torch>=2.0
rich>=13.0
matplotlib>=3.9
PyYAML>=5.0 